actors directory refactor
